+++ backend/core/src/seeds/listings/listing-default-seed.ts
@@ -15,7 +15,7 @@ import {
   getDefaultListingEvents,
   getDefaultProperty,
   getDefaultUnits,
-  getDisabilityOrMentalIlnessProgram,
+  getDisabilityOrMentalIllnessProgram,
   getDisplaceePreference,
   getHousingSituationProgram,
   getLiveWorkPreference,
@@ -113,14 +113,14 @@ export class ListingDefaultSeed {
       listingPreferences: [
         {
           preference: await this.preferencesRepository.findOneOrFail({
-            title: getLiveWorkPreference().title,
+            title: getLiveWorkPreference(alamedaJurisdiction.name).title,
           }),
           ordinal: 1,
           page: 1,
         },
         {
           preference: await this.preferencesRepository.findOneOrFail({
-            title: getDisplaceePreference().title,
+            title: getDisplaceePreference(alamedaJurisdiction.name).title,
           }),
           ordinal: 2,
           page: 1,
@@ -142,7 +142,7 @@ export class ListingDefaultSeed {
         },
         {
           program: await this.programsRepository.findOneOrFail({
-            title: getDisabilityOrMentalIlnessProgram().title,
+            title: getDisabilityOrMentalIllnessProgram().title,
           }),
           ordinal: 3,
         },
@@ -154,6 +154,7 @@ export class ListingDefaultSeed {
         },
       ],
       jurisdictionName: "Alameda",
+      jurisdiction: alamedaJurisdiction,
     }
 
     return await this.listingRepository.save(listingCreateDto)