+++ ui-components/__tests__/actions/ExpandableContent.test.tsx
@@ -1,13 +1,12 @@
 import React from "react"
 import { render, cleanup, fireEvent } from "@testing-library/react"
-
 import { ExpandableContent } from "../../src/actions/ExpandableContent"
 
 afterEach(cleanup)
 
 describe("<ExpandableContent>", () => {
   const INNER_CONTENT_CLASS = "inner-content"
-  const content = <div data-testid={INNER_CONTENT_CLASS}>Sample content</div>
+  const content = <div data-test-id={INNER_CONTENT_CLASS}>Sample content</div>
 
   it("renders content after button click", () => {
     const { getByText, queryByTestId } = render(<ExpandableContent>{content}</ExpandableContent>)